Refactor packages
Moved the content from fetcher package to fetcher.zalando to have the zalando API specific things in that package.
