--- products-ingester/src/main/scala/com/gsingharoy/productsingester/services/fetcher/RawBrand.scala
@@ -1,10 +1,0 @@
-package com.gsingharoy.productsingester.services.fetcher
-
-import io.circe.Decoder
-import io.circe.generic.semiauto
-
-private[fetcher] case class RawBrand(name: String)
-
-private[fetcher] object RawBrand {
-  implicit val decodeRawBrand: Decoder[RawBrand] = semiauto.deriveDecoder[RawBrand]
-}
