+++ products-ingester/src/main/scala/com/gsingharoy/productsingester/services/fetcher/zalando/RawProduct.scala
@@ -1,10 +1,10 @@
-package com.gsingharoy.productsingester.services.fetcher
+package com.gsingharoy.productsingester.services.fetcher.zalando
 
 import com.gsingharoy.productsingester.models.Product
 import io.circe.Decoder
 import io.circe.generic.semiauto
 
-private[fetcher] case class RawProduct(id: String,
+private[zalando] case class RawProduct(id: String,
                       modelId: String,
                       name: String,
                       brand: RawBrand) {
@@ -12,7 +12,7 @@ private[fetcher] case class RawProduct(id: String,
     Product(name = name, brand = brand.name, price = 0.0)
   }
 }
-private[fetcher] object RawProduct{
+private[zalando] object RawProduct{
   implicit val decodeRawProduct: Decoder[RawProduct] = semiauto.deriveDecoder[RawProduct]
 
 }
