+++ products-ingester/src/main/scala/com/gsingharoy/productsingester/services/fetcher/zalando/RawProductsResponse.scala
@@ -1,10 +1,10 @@
-package com.gsingharoy.productsingester.services.fetcher
+package com.gsingharoy.productsingester.services.fetcher.zalando
 
 import io.circe.Decoder
 import io.circe.generic.semiauto
 
 
-private[fetcher] case class RawProductsResponse(content: Seq[RawProduct],
+private[zalando] case class RawProductsResponse(content: Seq[RawProduct],
                             totalElements: Option[Int],
                             totalPages: Option[Int],
                             page: Option[Int],
@@ -12,6 +12,6 @@ private[fetcher] case class RawProductsResponse(content: Seq[RawProduct],
   lazy val products = content.map(_.toProduct)
 }
 
-private[fetcher] object RawProductsResponse{
+private[zalando] object RawProductsResponse{
   implicit val decodeProductsResponse: Decoder[RawProductsResponse] = semiauto.deriveDecoder[RawProductsResponse]
 }
